Moves functions to helper_functions
